+++ smx/sourcemod/handles.py
@@ -1,10 +1,12 @@
 from __future__ import annotations
 
-from typing import Any, Callable, Dict
+from typing import Callable, Dict, Generic, TypeVar
+
+V = TypeVar('V')
 
 
-class SourceModHandle:
-    def __init__(self, id: int, obj: Any, on_close: Callable[[], None] | None = None):
+class SourceModHandle(Generic[V]):
+    def __init__(self, id: int, obj: V, on_close: Callable[[], None] | None = None):
         self.id = id
         self.obj = obj
         self.on_close = on_close
@@ -26,8 +28,11 @@ class SourceModHandles:
     def __getitem__(self, handle_id: int):
         return self._handles[handle_id].obj
 
+    def get_raw(self, handle_id: int) -> SourceModHandle | None:
+        return self._handles.get(handle_id)
+
     def get(self, handle_id: int):
-        handle = self._handles.get(handle_id)
+        handle = self.get_raw(handle_id)
         if handle:
             return handle.obj