+++ smx/sourcemod/natives/__init__.py
@@ -1,8 +1,10 @@
 from __future__ import annotations
 
-from typing import Any, Callable, Dict, TYPE_CHECKING
+import re
+from pathlib import Path
+from typing import Any, Callable, Dict, List, Tuple, TYPE_CHECKING
 
-from smx.sourcemod.natives.base import MethodMap, native
+from smx.sourcemod.natives.base import MethodMap
 from smx.sourcemod.natives.console import ConsoleNatives
 from smx.sourcemod.natives.convars import ConVarNatives
 from smx.sourcemod.natives.files import FilesNatives
@@ -14,7 +16,20 @@ from smx.sourcemod.natives.string import StringNatives
 from smx.sourcemod.natives.timers import TimerNatives
 
 if TYPE_CHECKING:
+    from smx.runtime import SourcePawnPluginRuntime
     from smx.sourcemod.system import SourceModSystem
+
+
+RGX_NATIVE = re.compile(
+    r'''
+    ^native\s+
+    (?P<return_type>\S+)\s+
+    (?P<name>\w+)\s*
+    \((?P<params>.*)\)\s*
+    (?:;|$)
+    ''',
+    re.VERBOSE | re.MULTILINE,
+)
 
 
 class BaseSourceModNatives:
@@ -25,7 +40,7 @@ class BaseSourceModNatives:
         """
         self.sys = sys
         self.amx = self.sys.amx
-        self.runtime = self.amx.plugin.runtime
+        self.runtime: SourcePawnPluginRuntime = self.amx.plugin.runtime
 
     def get_native(self, qn: str) -> Callable[..., Any] | None:
         parts = qn.split('.', maxsplit=1)
@@ -51,6 +66,22 @@ class BaseSourceModNatives:
             obj = getattr(cls, name)
             if callable(obj) and getattr(obj, 'is_native', False):
                 natives[name] = obj
+        return natives
+
+    @staticmethod
+    def _read_sourcemod_natives(include_dir: str | Path | None = None) -> List[Tuple[Path, str]]:
+        """Get names of all natives in SourceMod's scripting/include dir"""
+        if include_dir is None:
+            from smx.compiler import INCLUDE_DIR
+            include_dir = INCLUDE_DIR
+        include_dir = Path(include_dir)
+
+        natives = []
+        for inc_path in include_dir.glob('**/*.inc'):
+            contents = inc_path.read_text()
+            for match in RGX_NATIVE.finditer(contents):
+                natives.append((inc_path, match.group('name')))
+
         return natives