+++ smx/sourcemod/natives/files.py
@@ -5,7 +5,8 @@ import io
 import os
 from enum import IntEnum
 
-from smx.sourcemod.natives.base import MethodMap, native, SourceModNativesMixin, WritableString
+from smx.sourcemod.handles import SourceModHandle
+from smx.sourcemod.natives.base import native, MethodMap, SourceModNativesMixin, WritableString
 from smx.sourcemod.printf import atcprintf
 
 
@@ -20,27 +21,27 @@ class File:
 
 
 class FileMethodMap(MethodMap):
-    @native('handle', 'writable_string')
-    def ReadLine(self, file: File, buf: WritableString):
-        line = file.fp.readline()
+    @native
+    def ReadLine(self, handle: SourceModHandle[File], buf: WritableString):
+        line = handle.obj.fp.readline()
         return buf.write(line, null_terminate=True)
 
-    @native('handle', 'cell', 'cell', 'cell')
-    def Read(self, file: File, items_addr: int, num_items: int, item_size: int):
+    @native
+    def Read(self, handle: SourceModHandle[File], items_addr: int, num_items: int, item_size: int):
         if item_size not in (1, 2, 4):
             # TODO(zk): native error
             raise ValueError('Unsupported item size %s' % item_size)
 
-        data = file.fp.read(num_items * item_size)
+        data = handle.obj.fp.read(num_items * item_size)
         items_read = len(data) // item_size
         data = data[:items_read * item_size]
 
         self.amx._writeheap(items_addr, ctypes.create_string_buffer(data))
         return items_read
 
-    @native('handle')
-    def EndOfFile(self, file: File):
-        return file.is_eof()
+    @native
+    def EndOfFile(self, handle: SourceModHandle[File]):
+        return handle.obj.is_eof()
 
 
 class PathType(IntEnum):
@@ -50,7 +51,7 @@ class PathType(IntEnum):
 class FilesNatives(SourceModNativesMixin):
     File = FileMethodMap()
 
-    @native('cell', 'writable_string', 'string', '...')
+    @native
     def BuildPath(self, path_type: PathType, buffer: WritableString, fmt: str, *args):
         if path_type != PathType.PATH_SM:
             raise ValueError('Unsupported path type %s' % path_type)
@@ -59,7 +60,7 @@ class FilesNatives(SourceModNativesMixin):
         path = str(self.runtime.root_path / suffix)
         return buffer.write(path, null_terminate=True)
 
-    @native('string', 'string', 'bool', 'string')
+    @native
     def OpenFile(self, file: str, mode: str, use_valve_fs: bool = False, valve_path_id: str = 'GAME'):
         if use_valve_fs:
             # TODO(zk)