+++ smx/sourcemod/natives/handles.py
@@ -2,7 +2,7 @@ from smx.sourcemod.natives.base import native, SourceModNativesMixin
 
 
 class HandlesNatives(SourceModNativesMixin):
-    @native('cell')
+    @native
     def CloseHandle(self, handle_id: int) -> None:
         # TODO(zk): run time error if handle is invalid
         self.sys.handles.close_handle(handle_id)