+++ smx/sourcemod/natives/shell.py
@@ -5,12 +5,12 @@ from typing import TYPE_CHECKING
 
 from smx.definitions import cell
 from smx.exceptions import SourcePawnRuntimeError
+from smx.interfaces import ParamCopyFlag, ParamStringFlag
 from smx.sourcemod.natives import FloatNatives
 from smx.sourcemod.natives.base import native, SourceModNativesMixin, WritableString
 
 if TYPE_CHECKING:
     from smx.runtime import PluginFunction
-    from smx.interfaces import ParamCopyFlag, ParamStringFlag
 
 
 def _format_float(value: float) -> str:
@@ -22,67 +22,67 @@ def _format_float(value: float) -> str:
 class ShellNatives(SourceModNativesMixin):
     """Natives used when running sourcepawn's test suite"""
 
-    @native('cell')
+    @native
     def printnum(self, value: int) -> int:
         s = f'{value}\n'
         self.runtime.printf(s)
         return len(s)
 
-    @native('cell')
+    @native
     def writenum(self, value: int) -> int:
         s = f'{value}'
         self.runtime.printf(s)
         return len(s)
 
-    @native('float')
+    @native
     def printfloat(self, value: float) -> int:
         s = f'{_format_float(value)}\n'
         self.runtime.printf(s)
         return len(s)
 
-    @native('float')
+    @native
     def writefloat(self, value: float) -> int:
         s = f'{_format_float(value)}'
         self.runtime.printf(s)
         return len(s)
 
-    @native('...')
+    @native
     def printnums(self, *addrs) -> int:
         nums = [self.amx._getheapcell(addr) for addr in addrs]
         self.runtime.printf(', '.join(str(x) for x in nums) + '\n')
         return 1
 
-    @native('string')
+    @native
     def print(self, value: str) -> int:
         self.runtime.printf(value)
         return len(value)
 
     @native
-    def dump_stack_trace(self):
+    def dump_stack_trace(self) -> None:
         self.runtime.printf(self.amx.dump_stack() + '\n')
 
     @native
-    def report_error(self):
+    def report_error(self) -> None:
         self.runtime.amx.report_error('What the crab?!')
 
     @native
-    def donothing(self):
+    def donothing(self) -> int:
         return 1
 
-    @native('cell')
+    @native
     def dynamic_native(self, arg: int) -> int:
         # XXX(zk): should this probe more?
         return arg
 
     # TODO(zk): add conveniences for by-ref params
-    @native('cell', 'cell', 'cell', 'cell')
+    @native
     def access_2d_array(self, array_addr: int, x: int, y: int, out_addr: int) -> int:
         row_addr = self.amx._getheapcell(array_addr + x * sizeof(cell))
         value = self.amx._getheapcell(row_addr + y * sizeof(cell))
         self.amx._writeheap(out_addr, cell(value))
         return 1
 
-    @native('function', 'cell')
+    @native
     def invoke(self, func: PluginFunction, count: int) -> int:
         for i in range(count):
             try:
@@ -93,7 +93,7 @@ class ShellNatives(SourceModNativesMixin):
 
         return 1
 
-    @native('function', 'cell')
+    @native
     def execute(self, func: PluginFunction, count: int) -> int:
         ok = 0
         for i in range(count):
@@ -107,7 +107,7 @@ class ShellNatives(SourceModNativesMixin):
 
         return ok
 
-    @native('function', 'writable_string', 'cell', 'cell')
+    @native
     def call_with_string(
         self,
         fn: PluginFunction,
@@ -127,27 +127,27 @@ class ShellNatives(SourceModNativesMixin):
         return call_rval.rval
 
     # TODO(zk): add some convenience wrappers around heap addresses
-    @native('cell', 'cell', 'cell', 'function')
+    @native
     def copy_2d_array_to_callback(
         self,
         flat_array_addr: int,
         length: int,
         stride: int,
         callback: PluginFunction,
-    ) -> None:
+    ) -> int:
         with self.amx._heap_scope():
             copied_array_addr = self.amx._heap_alloc_2d_array(length, stride, init_addr=flat_array_addr)
             callback(copied_array_addr, length, stride)
             return 0
 
-    @native('cell', 'cell')
+    @native
     def assert_eq(self, lhs: int, rhs: int) -> None:
         # TODO(zk): report error on comp failure
         assert lhs == rhs
 
     CloseHandle = donothing
 
-    @native('cell')
+    @native
     def __float_ctor(self, int_val: int):
         # XXX(zk): is this supposed to reinterpret a cell as a float? or turn an int into a float?
         return float(int_val)
@@ -159,35 +159,35 @@ class ShellNatives(SourceModNativesMixin):
     __float_mul = FloatNatives.FloatMul
     __float_div = FloatNatives.FloatDiv
 
-    @native('float', 'float')
+    @native
     def __float_mod(self, dividend: float, divisor: float) -> float:
         return dividend % divisor
 
-    @native('float', 'float')
+    @native
     def __float_gt(self, lhs: float, rhs: float) -> bool:
         return lhs > rhs
 
-    @native('float', 'float')
+    @native
     def __float_ge(self, lhs: float, rhs: float) -> bool:
         return lhs >= rhs
 
-    @native('float', 'float')
+    @native
     def __float_lt(self, lhs: float, rhs: float) -> bool:
         return lhs < rhs
 
-    @native('float', 'float')
+    @native
     def __float_le(self, lhs: float, rhs: float) -> bool:
         return lhs <= rhs
 
-    @native('float', 'float')
+    @native
     def __float_eq(self, lhs: float, rhs: float) -> bool:
         return lhs == rhs
 
-    @native('float', 'float')
+    @native
     def __float_ne(self, lhs: float, rhs: float) -> bool:
         return lhs != rhs
 
-    @native('float')
+    @native
     def __float_not(self, value: float) -> bool:
         if value == float('nan'):
             return True