+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  - Add `GetPluginFilename` native
 
+### Changed
+ - Switch `@native` decorator to interpret param types from typing annotations
+
 
 ## [0.4.0] — 2023-03-02
 ### Added